+++ src/main/java/com/somemore/auth/jwt/filter/UserRole.java
@@ -1,4 +1,4 @@
-package com.somemore.auth;
+package com.somemore.auth.jwt.filter;
 
 public enum UserRole {
     VOLUNTEER,
