+++ src/main/java/com/somemore/volunteer/usecase/FindVolunteerIdUseCase.java
@@ -1,4 +1,4 @@
-package com.somemore.volunteer.usecase.query;
+package com.somemore.volunteer.usecase;
 
 import com.somemore.volunteer.dto.response.VolunteerForCommunityResponseDto;
